pyproject.toml
[project]
name = "pdf-to-png"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "pymupdf>=1.25.5",
    "pytest>=9.0.2",
    "reportlab>=4.4.9",
]
